// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/phrase-occurrence-page-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PhraseOccurrenceDto } from '../models/phrase-occurrence-dto';
import { PhraseRepetitionDetailDto } from '../models/phrase-repetition-detail-dto';
export interface PhraseOccurrencePageResponse {
  activeBuildId: string;
  items: Array<PhraseOccurrenceDto>;
  page: number;
  pageSize: number;
  phrase: PhraseRepetitionDetailDto;
  totalCount: number;
}