// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/phrase-repetitions-page-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PhraseRepetitionListItemDto } from '../models/phrase-repetition-list-item-dto';
export interface PhraseRepetitionsPageResponse {
  activeBuildId: string;
  items: Array<PhraseRepetitionListItemDto>;
  mode: string;
  page: number;
  pageSize: number;
  sort: string;
  totalCount: number;
  wordCount: number;
}